Github issue #9
https://github.com/whosgonna/Zabbix-Tiny.pm/issues/9

While most Zabbix API methods take alist of key=>value pair params, some
(such as the *.delete methods) only take an array of values.  The
existing code prevented this from working because it blindly assumed a
hash would be used.

The do method was altered to check for a single argument.  This single
argumenet can be either an arrayref (for the aformentioned scenario
where the params are just a list of values), or a hashref of all values
for the params.

It is recommended to use always use this way of passing the arguments,
however overloading the method will cause it to revert to the previous
behavior, allowing for backwards compatability.

Old code can be updated by adding curly brackets around the 2nd through
n-th aguments.  Thus:

$zabbix->do( 'host.get', groupids => 12345, limit => 1 );

should be expressed as:

$zabbix->do( 'host.get', { groupids => 12345, limit => 1 } );
                         ^                               ^