GitHub issue12:  error not displayed when connection to the db fails
Some errors from the Zabbix API put the interesting data into 'data'
field of the error response, which is how Zabbix::Tiny was originally
written to provide the error message to users.  Some other error
responses (Zabbix not connected to it's database) have the interesting
data in the 'message' field of the error response.  Zabbix::Tiny will
now provide both messages on failure.