Added tests to t/02_apiinfo.version.t
Now confirm that the version can be retrieved without login (will fail
with the auth parameter sent), then confirm that the object has not
logged in. Next, login and confirm the auth id.  Finally, request the
version number again to confirm that the auth is not sent in post-login
requests for the version.